chore(util): moved CalendarDate to utils
